version 8
Se corrigen errores y se modifica la url para tener relación con el resto de APIS

+++ src/main/java/com/SAFE_Rescue/API_Administrador/service/BomberoService.java
@@ -127,6 +127,10 @@ public class BomberoService {
 
     public void delete(long id){
         try {
+            if (!bomberoRepository.existsById(id)) {
+                throw new NoSuchElementException("Bombero no encontrado");
+            }
+
             bomberoRepository.deleteById(id);
 
         } catch (Exception e) {
